Another 64->32 bit warning from the Windows compiler. Review URL: http://codereview.chromium.org/8503010
git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@9910 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/incremental-marking.cc
@@ -862,9 +862,9 @@ void IncrementalMarking::Step(intptr_t allocated_bytes) {
       }
     } else {
       allocation_marking_factor_ += kAllocationMarkingFactorSpeedup;
-      allocation_marking_factor_ =
+      allocation_marking_factor_ = static_cast<int>(
           Min(kMaxAllocationMarkingFactor,
-              static_cast<intptr_t>(allocation_marking_factor_ * 1.3));
+              static_cast<intptr_t>(allocation_marking_factor_ * 1.3)));
       if (FLAG_trace_gc) {
         PrintF("Marking speed increased to %d\n", allocation_marking_factor_);
       }
